Extract object-links
