+++ app/scripts/views/object.js
@@ -1,33 +1,10 @@
 import React, {Component, PropTypes} from 'react'
 import {Link} from 'react-router'
-import {Table} from 'react-bootstrap'
 
 import i18n from '../utils/i18n'
 import Icon from './icon'
 import RawData from './object/raw-data'
-import ObjectLink from './object/object-link'
-
-const ObjectLinks = ({path, links}) => {
-  return (
-    <div>
-    <li className='list-group-item'>
-      <strong>{i18n.t('Object links')}</strong>
-    </li>
-    <Table responsive className='filelist'>
-      <thead>
-        <tr>
-          <th>{i18n.t('Name')}</th>
-          <th>{i18n.t('Hash')}</th>
-          <th>{i18n.t('Size')}</th>
-        </tr>
-      </thead>
-      <tbody>
-        {links.map(link => <ObjectLink key={link.Hash} path={path} link={link} />)}
-      </tbody>
-    </Table>
-    </div>
-  )
-}
+import ObjectLinks from './object/object-links'
 
 const ParentLink = ({parent}) => {
   if (!parent) return <span></span>
